+++ spec/shared_examples.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require 'shared_examples/package'
+require 'shared_examples/reload'
+require 'shared_examples/repository'
+require 'shared_examples/service'